Move common test fixtures to a shared file, conftest.py
